// src/lib/types/letterboxd.ts
export interface Image {
    sizes: ImageSize[];
}

export interface ImageSize {
    height: number;
    url: string;
    width: number;
}
// ...
export interface FilmSummary {
    /**
     * `true` if the film is in TMDb's 'Adult' category.
     */
    adult: boolean;

    /**
     * The film's unobfuscated poster image (2:3 aspect ratio in multiple sizes), only populated if
     * the `adult` flag is `true`, may contain adult content.
     */
    adultPoster: Image;

    /**
     * The other names by which the film is known (including alternative titles and/or foreign translations).
     */
    alternativeNames: string[];

    /**
     * The LID of the collection containing this film.
     */
    filmCollectionId: string;

    /**
     * The LID of the film.
     */
    id: string;

    /**
     * A list of relevant URLs for this entity, on Letterboxd and external sites.
     */
    links: Link[];

    /**
     * The title of the film.
     */
    name: string;

    /**
     * The original title of the film, if it was first released with a non-English title.
     */
    originalName: string;

    /**
     * The film's poster image (2:3 aspect ratio in multiple sizes). Will contain only a single
     * obfuscated image if the `adult` flag is `true`.
     */
    poster: Image;

    /**
     * The year in which the film was first released.
     */
    releaseYear: number;

    /**
     * The film's duration (in minutes).
     */
    runTime: number;
}

export interface Link {
    /**
     * The object ID for the linked entity on the destination site.
     */
    id: string;

    /**
     * Denotes which site the link is for.
     */
    type:
    | 'facebook'
    | 'gwi'
    | 'imdb'
    | 'instagram'
    | 'justwatch'
    | 'letterboxd'
    | 'ticket'
    | 'tmdb'
    | 'twitter'
    | 'youtube';

    /**
     * The fully qualified URL on the destination site.
     */
    url: string;
}